Se comenzo a pasar de entero a decimal con el fin de agregar decimales a las cantidades, totales, subtotales, precio entre otros

+++ src/database/migrations/17-create-detalle-ajuste.js
@@ -26,11 +26,11 @@ module.exports = {
         }
       },
       cantidad_antes: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(15, 3),
         allowNull: false
       },
       cantidad_ahora: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(15, 3),
         allowNull: false
       }
